+++ main.py
@@ -10,7 +10,7 @@ app = create_app(config)
 
 if __name__ == "__main__":
 
-    host = os.getenv("HOST", "127.0.0.1")
+    host = os.getenv("HOST", "0.0.0.0")
     port = int(os.getenv("PORT", "8080"))
 
     uvicorn.run(app, host=host, port=port)
